Fixed umlaut handling in links without caption
In tinymce, innerhtml comes with html entities. So umlauts are entity
encoded. When caption and link target are identical, the editor should not
produce a caption in wikitext. Due to different encoding, this was the
case until now. Now, encoding is the same and thus identical caption and
target are now recognized even if there are umlauts.

Change-Id: I00457510708afb2e5b8273faf836960abea36e76