1072 [BS224] PermissionManager: LockMode & pm-settings.php Backup
I added extra checks if a new configuration file has to be built. Lockmode now writes a new pm-settings.php, only if any of its needed settings are not set.
Additionally, I also added a session flag to short-circuit the whole checking process if the process was running once in the current session AND if the state of Lockmode (active|inactive) didn't changed.

I tried to reformat the file closely to the styleguide.

Change-Id: I83ecc91d58e1f3efba2451f5945b2ec999f9af2c