Fixed skin usage for anonymous users
Flexiskin uses an internal API call in order to add specific additional
skin elements. The API was set to require read rights. This lead to a
MWException if Flexiskin was enabled and no user logged in in wikis that
do prevent read access for anonymous users. Fixed now by unsetting the
required rights.

Change-Id: I58012841c8a2797c5f4821b4f59d6cec3c977cc8