GroupManager: Raised group name character limit
The limit for characters of a group name was raised to 255 in the
MediaWiki db schema, but the NamespaceManager still had it at 16. The
limit is now set to 255 accordingly.

PatchSet 2: updated i18n manually, as this branch is not maintained by
translatewiki.net

Change-Id: I3b8ce9c2b62a61aec5547bca88a3fc07e16c1b34
(cherry picked from commit 5b55510a30717f060ad60b23cb8a8da8c4dbf86e)