NamespaceManager: Systemnamespaces can no longer be deleted
When you try to delete a system namespace, the NamespaceManager will
reject this. This check somehow got lost in recent development, leading
to the possibility to delete all pages, e.g. in main namespace, with one
click. The check is now restored, plus checks are also done on the
server side now.

PatchSet 2: implemented CR

Change-Id: I2cce6ea0612b5001800fde65297760c505df6aad