Blog: repace action=blog with special page
Option action=blog was replaced with Special:Blog, which essentially
covers the same functionality. In addition, you can use a namespace
name or a category as a selection parameter. The old action was
deprecated, but redirects to the special page for legacy reasons.

BlueSpice internal: ERM3719

Patch Set 2: Implemented changes per Robert Vogel

Change-Id: I59e5b9c2853eb0b6829550880970417320228de3