Blog: Remove custom caching
There was a custom caching in blog which lead to the effect that blog
articles were only published after one day. Caching makes no sense in
the current implementation, as any change on the wiki must lead to cache
invalidation. On top of this, timestamps are also generated server side,
so cache must be invalidated every 5 minutes or so. Overall, caching
strategy for blog needs to be retought, should performance issues arise.

Change-Id: I4a403205453125b2466e2eae3b7503ef2ea0b63e