Fix for multiple extended search result divs
When additional filters are applied to ExtendedSearch results, <div> with
results is duplicated instead of just being updated. When filters are applied
and new AJAX request is made, old <div> should be removed and new one
should be created.
jQuery.remove() removes only 1 element, in this case there are multiple
matches, which caused results <div> to be left not-deleted, so therefore
duplicated on page. This iterates over all elements are removes each
separately

This needs to be cherry-picked to REL1_27 and master

Change-Id: Idf4896642f52593ca4f818f99c3bc5e71f122713