Refactoring of ListView on Statistics
The list view was created by returning HTML table and passing it
to Ext.ux.grid.TransformGrid. Now implemened by regular Ext.grid.Panel

Added field type, to enable correct sorting of numeric columns.

Added paging of the list

This removed running of a hook BSStatisticsBeforeDrawTable. I am not sure
what to do with it. No one attaches to this hook, and i see no
documentation of it. Should i re-implement it as a different hook so
content of data sent to grid can be changed?

Change-Id: I9ce4eb0129265dc358cb70b72fa260ec97d12028