Fix for error on inserting wrong link in InsertLink dialog
When user enters a link with section identifier (#) or ?, & and clicks OK,
dialog crashes, because no valid title is being selected. This is a simple
fix to prevent such unlikely events.

I tried to make it actualy work, to be able to get valid title even from
link that contains wrong characters, but it will take some time. If anyone
thinks that it would be good to do, I will be happy to refactor it!

This also prevents errors on accidentaly clicking OK without link set

Needs cherry-picking to REL1_27 and master

Change-Id: Ia120b14c3c1d715fbc10e0326ee73fd46a9bbbe7