Fix for broken dialog on PermissionManager TemplateEditor
On tree panel selection change (beforeselect), if there are changes in
current record, a dialog asking wether to save changes is displayed. If
user says Yes, saveTemplate is called and state is changed to 'clean' (no
changes). Problem is that this setting to clean happens in .done method of
ajax call which comes late, before it gets the chance to set the state to
clean, beforeselect event is fired again, and it tries to dispay the
dialog again, but this time it failes, causing broken dialog.

Added also _closeDialog var, to distinguish when dialog should be closed
(Save button clicked) and when it should remain open (saving after
changing selection)

ERM: #3812

Needs cherry-picking to REL1_27 and REL1_23

Change-Id: I975be9eb6ced6571078dc04a62664be7d83eb784