Fix for checkboxes removing when alone in the line
When a checkbox (bs:checklist) is alone in the line it gets removed due to
tinyMCE cleaning that removes empty elements, in this case empty <button>
is created. This alters this behaviour by adding a space char inside
button before setting content and removes it when content is set.

Needs cherry-picking to REL1_27 and master

Change-Id: Iee42855360ee5ac3e38084fc2330847da2360623