VisualEditor: introducing DOM preprocessor
A DOM preprocessor is introduced which currently only replaces certain
spans (underline and strike). This is neccessary because it is the
correct way to handle nested spans, where regexs would fail. In the wake
of this change, handling of external links was improved, as the html
produced was a mess.

Patch Set 2: Implemented CR

Change-Id: I6a63524576ebc0869bb77c9621106d563625f390