Register MagicWords as objects not arrays
If new magic words are registered using array syntax it will crash
filtering, must use stdClass

Filtering of insertMagic store is not done anywhere now, but if it was it
would crash

Now even if MagicWord are registered like an array,
they will be converted to stdClass object.

Added examples and helplink to all instances, even if they are not used,
for conformity.

@PWirth: previewable was in InsertMagic extension, not sure where it is
used, but i didnt want to remove it, in case there is actually some need
for it

Needs cherry-picking to REL1_27 and REL1_23

Change-Id: I73289a982d907b16f6a019b7d68ce5ea7418071c