Workaround for InsertLink in IE
In IE, when replacing a text with a link in VE, it crashes.

This solution doesnt use editor.insertContent but handles dom elements.
If you select a word in a paragraph, getNode() and getSel() return entire
paragraf. After trying many solutions, only this one works, it gets node
text, replaces selected word(s) with anchor tag, creates new node, and
replaces it with the old one.

** Same problem exists with InsertFile and InsertMagic but I didnt want to
change them until this is accepted.

Needs cherry-picking to REL1_27 and REL1_23

Change-Id: Ie2f4dcd8fded5f3a55cc8b902406ee355c78b6bf
ERM: #5653