Avatars: Fix for image width
If user uploads own image to be profile image over Avatars dialog
image width (and possibly height) will be forced to be whatever is set as
default image width/height. This causes some images to lose their aspect
ratio.

I didnt want to change generateAvatar function to include width/height of
actual image in response, so i made the change on creating image element
itself. Because of possible unforseen consenquences, image width/height is
only changed if its smaller than the values in existing params, because in
that case it doesnt make sense to enlarge image beyond its actual size

Even if user uploads .jpg image as a profile picture, it will be  converted
to .png

ERM: #6046

Needs cherry-picking to REL1_27 and REL1_23

Change-Id: I9ca3cf0cab3d38128bb56d47494e8407b1c99dbd