ExtendedSearch: Fix for sending multiple namespace params to Solr
When term is typed into autocomplete box search query is created with
default namespaces to be searched. But when term with namespace already in
is typed in 'namespace' key is added again, causing missed matches.

Now default namespaces are added only if no explicit NS is detected

Needs cherry-picking to REL1_27 and REL1_23

Change-Id: I9507a8e0eaac1e63b9f5ddfe0fcf9cad60f706d6
ERM: #6247