PermissionManager: restore global state after API call
getPermissionArray() messes with permission related globals without
resetting them at the end of the method. In the case of a single API
call, this is no problem, but could lead to issues with internal
fake API calls. So the global context is restored at the end of the
method now.

Patch Set 2: Implemented CR by Robert

Change-Id: Ic14b8a252023b3cac533ec48072a4dee430082da