Review: Add custom notes
Customer complained that its not clear that the comment will be deleted
when delegating the review and wanted a message saying that.

In order to implement such a message we needed a method that will add
notes to review.

I meant this to display message as received from caller, and caller should
be, depending on the content, responsible for possible escaping, is this
ok?

Removed some trailing whitespaces from before

Change-Id: Ib347fc0d41a2b8a8df91e97af28613c90c2da05a
ERM: #6182
Needs cherry-picking to REL1_27 and REL1_23