UEModulePDF: Change tag filename is retrieved from
When getting files from the page, data attribute data-bs-filename was
checked. This attribute exists only on Media tags and does not contain
possible extra namespace file is in. This caused files that are not only
in NS_FILE, to be missed when creating PDF, because Title object was
created from non-ns-prefixed name.

Only attribute that contains full name with NS prefix is data-bs-title.
I have found no disadvantages in using it instead of the old tag, it
exists on every file tag, contains the same name as data-bs-filename with
addition of namespace, which is no problem when creating Title object from
it. When using it files in other namespaces will be detected.

ERM: #4275

Needs cherry-picking to REL1_27 and REL1_23

Change-Id: Id16373130c5bf6d56f09dec3d8f0609f7749542a