UEModulePDF: Fix for tables being too wide
In tables with many columns it can happen that the table overflows the
width of the size in PDF export.

Since table-layout must be set, it cannot be done by simple setting
"width" attribute of table, but in this, a little hacky way, to alter
"style" tag.

It can also be done over css class, which would more elegantly be added to
tables in php, but this might not work, in case wiki uses custom skin,
which does not have class defined.

Change-Id: I85f3078a68c22b4b8a187842a3548132d58b9a6f
ERM: #3591