VisualEditor: Fix for broken syntax by "link in caption"
If wiki text contains an image that has another link in its caption
and also the same link from the caption is used again on page, syntax
breaks.

Its because when parsing the second link, it finds it also in the image
caption and replaces it with html.

Since thumbs and framed images are not parsed with a frame anyway, caption
is not important in VE, so i used just escaping so that it doesnt break

Change-Id: I972cac06b3f29af6e216ae314dfafa4949821270
ERM: #7784
(cherry picked from commit f4d16e63cb349037bec2ecf4cbb1cd04f25c1eff)