Fix for NSFR files
When moving file, old file should be removed from index. This File object
is created by calling LocalFile::newFromTitle which always returns
LocalFile, and not NSLocalFile in case NSFR is installed.
This causes getPath() to return filename with NS prefix, making it
impossible to remove from index.

It causes old files to remain in index as redlinks

Change-Id: I08a99a9d0ef0c2181e4a5fcddb1d06e5df448a8f
ERM: #8146