Fix for NSFR files
When moving file, old file should be removed from index. This File object
is created by calling LocalFile::newFromTitle which always returns
LocalFile, and not NSLocalFile in case NSFR is installed.
This causes getPath() to return filename with NS prefix, making it
impossible to remove from index.

It causes old files to remain in index as redlinks

Change-Id: Ia515729f4310e09ba3970381a5f4d6cc84612a1e
ERM: #8146