UEModulePDF/FileResolver: Consider file timestamp
In the past, the actual target URL of an anchor/image element was analyzed to
calculate the filesystem path. That way URLs pointing to 'archive' got
resolved correctly.

Then BSF added 'data-bs-*'-attributes to all kinds of links (e.g. with the
original file name). By that it got possible to ask the FileRepo for the
filesystem path, which is much more reliable. Unfortunately the information
whether or not the file was in the 'archive' got lost.

Now the file timestamp is being evaluated again, this time using a new
'data-bs-*'-attribute.

This change needs https://gerrit.wikimedia.org/r/#/c/423837 to work.

NEEDS CHERRY-PICK to REL1_27_dev

Change-Id: I46fbdabe5fda35bb32524bdeadd9f0b0f8848a54