Enable indexing of pages without sections
When a page had no sections the indexing was aborted.
When the page content could not be retrieved the indexing was aborted.

By this change the page get's indexed either way. No proper debug log
has been implemented as this is a legacy branch and ExtendedSearch has been
completely rewritten in the new version.

ERM10784

Change-Id: I5c451012665be3a2b4d9d8c23cba364e6564c49e