Add uniformity to tags
If tags have different name than the tag (blog <bs:blog>), InsertMagic
cannot recognize it when editing, which makes many of them uneditable.

We could use "id" instead of changing the name, but that will break vars
and switches -.-

If we want to have more user-friendly names displayed in the InsertMagic
grid, we should introduce some displayName key.

Change-Id: I450def7d51119fcd8a9b9891a404c0ac997f7ee0
ERM: #7289