Save Userpreferences with Caching
When cache was set to memcached, saving bluespice user preferences often lead
to overwritting every bluespice user option with the default.

Cause for wrong behaviour was not found;
Also: Setting for first time can still be not working.
Also: Restoring to default settings only works, when not in bluespice tab in settings

Only REL1_27

Change-Id: I99f8503e64c1f56110527e9f76c87116e5de1007